rm wrapped translator
